Extracted utils to a separted file
